requirements.txt
awscli
ruff
ipython
jupyterlab
loguru
matplotlib
mkdocs
notebook
numpy
pandas
pip
pytest
pytest-mock
pytest-cov
pytest-xdist
requests-mock
moto[s3]
evidently
python-dotenv
scikit-learn
tqdm
typer
mlflow
xgboost
prefect
prefect-aws
boto3
-e .
